Implemented loading of complex objects

+++ app/src/main/java/org/ccomp/data/domain/feed/Feed.java
@@ -16,11 +16,11 @@ public class Feed {
         return title;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -53,7 +53,7 @@ public class Feed {
     }
 
     @PrimaryKey(autoGenerate = true)
-    private int id;
+    private long id;
 
     @ColumnInfo(name = "title")
     private String title;
